+++ modules/core/server/models/content/querybuilder.js
@@ -219,9 +219,6 @@ module.exports = require(LACKEY_PATH)
 
                         query += ' LIMIT ' + num_limit;
 
-                        console.log(countQuery);
-                        console.log(query);
-
                         return Promise.all([
                             SCli.sql(knex.raw(countQuery)).then((r) => r.rows),
                             SCli.sql(knex.raw(query)).then((r) => r.rows)